+++ api_docs/submitClearingTransaction_ts.ts
@@ -2,7 +2,7 @@ import { ActionType } from '../src/types/api/actions';
 
 import type { ClearingTransactionPayload, ClearingTransactionResponse, CTAtomic } from '../src/types/api/clearingTransaction';
 import type { CTActionOrigin } from '../src/types/api/common';
-import type { CTAtomicAction, NftsMintItemArgs, NftsSetItemMetadataArgs } from '../src/txwrapper/calls';
+import type { CTAtomicAction, NftsMintItemArgs, NftsSetItemMetadataArgs } from '../src/txwrapper';
 
 // API endpoint
 const CLEARING_TRANSACTION_ENDPOINT = 'localhost:8080/submit/clearing_transaction';
