--- src/txwrapper/appTransactions/index.ts
@@ -1,1 +1,0 @@
-export { ClearingTransactionArgs, clearingTransaction } from './clearingTransaction';
