+++ src/txwrapper/index.ts
@@ -1,2 +1,1 @@
-export * from './appTransactions';
-export * from './calls';
+export * from './generated/calls';
